Change response time to fire non stop requests

+++ worker/app.py
@@ -1,8 +1,10 @@
 import requests
+import time
 import boto3
 from ec2_metadata import ec2_metadata
 import os
 from const import LB_PUBLIC_IP, TIME_OUT, PORT, HARAKIRI
+from datetime import datetime
 
 
 def work(buffer, iterations):
@@ -14,16 +16,22 @@ def work(buffer, iterations):
 
 
 def main():
+    start_time = datetime.utcnow()
     while True:
-        request = requests.get(f'http://{LB_PUBLIC_IP}:{PORT}/get_job', timeout=TIME_OUT)
+        dif = datetime.utcnow() - start_time
+        request = requests.get(f'http://{LB_PUBLIC_IP}:{PORT}/get_job')
         if request:
             job = request.json()
             res = work(job["file"], job["iterations"])
             requests.put(f"http://{LB_PUBLIC_IP}:{PORT}/return_result", json={"job_id": job["job_id"],
                                                                               "result": res})
+            start_time = datetime.utcnow()
+
         else:
-            if HARAKIRI:
+            if dif.seconds > 10 and HARAKIRI:
                 os.system('sudo shutdown -h now')
+
+            time.sleep(1)
 
 
 if __name__ == '__main__':
